update application files

+++ src/main/java/com/training/licenselifecycletracker/service/DeviceServiceImpl.java
@@ -1,13 +1,130 @@
 package com.training.licenselifecycletracker.service;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
+//import java.util.List;
+//import java.util.Optional;
+//import java.util.stream.Collectors;
+//
+//import org.modelmapper.ModelMapper;
+//import org.springframework.beans.factory.annotation.Autowired;
+//import org.springframework.stereotype.Service;
+//import org.springframework.transaction.annotation.Transactional;
+//
+//import com.training.licenselifecycletracker.dto.DeviceDTO;
+//import com.training.licenselifecycletracker.dto.LifecycleEventDTO;
+//import com.training.licenselifecycletracker.dto.SoftwareDTO;
+//import com.training.licenselifecycletracker.entities.Device;
+//import com.training.licenselifecycletracker.entities.LifecycleEvent;
+//import com.training.licenselifecycletracker.entities.Software;
+//import com.training.licenselifecycletracker.repositories.DeviceRepository;
+//import com.training.licenselifecycletracker.repositories.LifecycleEventRepository;
+//import com.training.licenselifecycletracker.repositories.SoftwareRepository;
+//
+//@Service
+//public class DeviceServiceImpl implements DeviceService {
+//
+//    private final DeviceRepository deviceRepository;
+//    private final SoftwareRepository softwareRepository;
+//    private final LifecycleEventRepository lifecycleEventRepository;
+//    private final ModelMapper modelMapper;
+//
+//    @Autowired
+//    public DeviceServiceImpl(DeviceRepository deviceRepository,
+//                             SoftwareRepository softwareRepository,
+//                             LifecycleEventRepository lifecycleEventRepository,
+//                             ModelMapper modelMapper) {
+//        this.deviceRepository = deviceRepository;
+//        this.softwareRepository = softwareRepository;
+//        this.lifecycleEventRepository = lifecycleEventRepository;
+//        this.modelMapper = modelMapper;
+//    }
+//
+//    @Override
+//    @Transactional
+//    public DeviceDTO addDevice(DeviceDTO deviceDTO) {
+//        Device device = modelMapper.map(deviceDTO, Device.class);
+//        Device savedDevice = deviceRepository.save(device);
+//        return modelMapper.map(savedDevice, DeviceDTO.class);
+//    }
+//    
+//    
+//    @Override
+//    @Transactional
+//    public DeviceDTO updateDevice(DeviceDTO deviceDTO) {
+//        Device deviceToUpdate = deviceRepository.findById(deviceDTO.getDeviceId())
+//                .orElseThrow(() -> new RuntimeException("Device not found with id: " + deviceDTO.getDeviceId()));
+//
+//        // Update existing device entity
+//        Device device= new Device();
+//        modelMapper.map(deviceDTO, device);
+//        System.out.println(device);
+//        device.getLifecycleEvent().setDevice(device);
+//
+//        // Save the updated device
+//        deviceRepository.save(device);
+//        DeviceDTO upDto = modelMapper.map(device, DeviceDTO.class);
+//        System.out.println(upDto);
+//        return upDto ;
+//    }
+//
+//    
+//
+//    @Override
+//    @Transactional(readOnly = true)
+//    public DeviceDTO getDeviceById(Integer deviceId) {
+//        Optional<Device> optionalDevice = deviceRepository.findById(deviceId);
+//        return optionalDevice.map(device -> modelMapper.map(device, DeviceDTO.class)).orElse(null);
+//    }
+//
+//    @Override
+//    @Transactional(readOnly = true)
+//    public List<DeviceDTO> getAllDevices() {
+//        List<Device> devices = (List<Device>) deviceRepository.findAll();
+//        return devices.stream()
+//                .map(device -> modelMapper.map(device, DeviceDTO.class))
+//                .collect(Collectors.toList());
+//    }
+//
+//    @Override
+//    @Transactional
+//    public void deleteDevice(Integer deviceId) {
+//        deviceRepository.deleteById(deviceId);
+//    }
+//
+//    @Override
+//    @Transactional
+//    public DeviceDTO addSoftwareToDevice(Integer deviceId, SoftwareDTO softwareDTO) {
+//        Optional<Device> optionalDevice = deviceRepository.findById(deviceId);
+//        if (optionalDevice.isPresent()) {
+//            Device device = optionalDevice.get();
+//            Software software = modelMapper.map(softwareDTO, Software.class);
+//            software.setDevice(device);
+//            Software savedSoftware = softwareRepository.save(software);
+//            device.getSoftwareList().add(savedSoftware);
+//            deviceRepository.save(device);
+//            return modelMapper.map(device, DeviceDTO.class);
+//        } else {
+//            throw new RuntimeException("Device not found with id: " + deviceId);
+//        }
+//    }
+//
+//    @Override
+//    @Transactional
+//    public DeviceDTO addLifecycleEventToDevice(Integer deviceId, LifecycleEventDTO lifecycleEventDTO) {
+//        Optional<Device> optionalDevice = deviceRepository.findById(deviceId);
+//        if (optionalDevice.isPresent()) {
+//            Device device = optionalDevice.get();
+//            LifecycleEvent lifecycleEvent = modelMapper.map(lifecycleEventDTO, LifecycleEvent.class);
+//            lifecycleEvent.setDevice(device);
+//            LifecycleEvent savedLifecycleEvent = lifecycleEventRepository.save(lifecycleEvent);
+//            device.setLifecycleEvent(savedLifecycleEvent);
+//            deviceRepository.save(device);
+//            return modelMapper.map(device, DeviceDTO.class);
+//        } else {
+//            throw new RuntimeException("Device not found with id: " + deviceId);
+//        }
+//    }
+//}
 
-import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import com.training.licenselifecycletracker.dto.DeviceDTO;
 import com.training.licenselifecycletracker.dto.LifecycleEventDTO;
@@ -15,28 +132,38 @@ import com.training.licenselifecycletracker.dto.SoftwareDTO;
 import com.training.licenselifecycletracker.entities.Device;
 import com.training.licenselifecycletracker.entities.LifecycleEvent;
 import com.training.licenselifecycletracker.entities.Software;
+import com.training.licenselifecycletracker.exceptions.DeviceNotFoundException;
 import com.training.licenselifecycletracker.repositories.DeviceRepository;
 import com.training.licenselifecycletracker.repositories.LifecycleEventRepository;
 import com.training.licenselifecycletracker.repositories.SoftwareRepository;
+import org.modelmapper.ModelMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class DeviceServiceImpl implements DeviceService {
 
-    private final DeviceRepository deviceRepository;
-    private final SoftwareRepository softwareRepository;
-    private final LifecycleEventRepository lifecycleEventRepository;
-    private final ModelMapper modelMapper;
 
+    
     @Autowired
-    public DeviceServiceImpl(DeviceRepository deviceRepository,
-                             SoftwareRepository softwareRepository,
-                             LifecycleEventRepository lifecycleEventRepository,
-                             ModelMapper modelMapper) {
-        this.deviceRepository = deviceRepository;
-        this.softwareRepository = softwareRepository;
-        this.lifecycleEventRepository = lifecycleEventRepository;
-        this.modelMapper = modelMapper;
-    }
+    DeviceRepository deviceRepository;
+    
+    @Autowired
+    SoftwareRepository softwareRepository;
+    
+    @Autowired
+    LifecycleEventRepository lifecycleEventRepository;
+    
+    @Autowired
+    ModelMapper modelMapper;
+    
+    
+    
 
     @Override
     @Transactional
@@ -49,15 +176,19 @@ public class DeviceServiceImpl implements DeviceService {
     @Override
     @Transactional
     public DeviceDTO updateDevice(DeviceDTO deviceDTO) {
-        Optional<Device> optionalDevice = deviceRepository.findById(deviceDTO.getDeviceId());
-        if (optionalDevice.isPresent()) {
-            Device deviceToUpdate = optionalDevice.get();
-            modelMapper.map(deviceDTO, deviceToUpdate); // Update existing device entity
-            deviceRepository.save(deviceToUpdate);
-            return modelMapper.map(deviceToUpdate, DeviceDTO.class);
-        } else {
-            throw new RuntimeException("Device not found with id: " + deviceDTO.getDeviceId());
+        Device deviceToUpdate = deviceRepository.findById(deviceDTO.getDeviceId())
+                .orElseThrow(() -> new RuntimeException("Device not found with id: " + deviceDTO.getDeviceId()));
+
+        modelMapper.map(deviceDTO, deviceToUpdate);
+
+        // Ensure lifecycle event association is correctly set
+        if (deviceToUpdate.getLifecycleEvent() != null) {
+            deviceToUpdate.getLifecycleEvent().setDevice(deviceToUpdate);
         }
+
+        deviceRepository.save(deviceToUpdate);
+
+        return modelMapper.map(deviceToUpdate, DeviceDTO.class);
     }
 
     @Override
@@ -76,11 +207,23 @@ public class DeviceServiceImpl implements DeviceService {
                 .collect(Collectors.toList());
     }
 
+//    @Override
+    
+//    public void deleteDevice(Integer deviceId) {
+//        deviceRepository.deleteById(deviceId);
+//    }
+    
     @Override
     @Transactional
-    public void deleteDevice(Integer deviceId) {
-        deviceRepository.deleteById(deviceId);
+    public void deleteDevice(Integer deviceId)  {
+        Optional<Device> device = deviceRepository.findById(deviceId);
+        if (device.isPresent()) {
+        	deviceRepository.deleteById(deviceId);
+           
+        } 
     }
+    
+    
 
     @Override
     @Transactional
